Remove binaryITCase
